# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  username: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface SocialLinks {
  /** @nullable */
  github?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface Profile {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  domain?: string | null;
  socialLinks?: SocialLinks;
}

export interface Project {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  techStack?: string[];
  /** @nullable */
  repoLink?: string | null;
  /** @nullable */
  liveLink?: string | null;
  /** @nullable */
  screenshotUrl?: string | null;
}

export interface Skills {
  frontend?: string[];
  backend?: string[];
  devops?: string[];
  other?: string[];
}

export interface PortfolioUser {
  id: number;
  username: string;
  email: string;
  templateId: string;
  profile?: Profile;
  projects?: Project[];
  skills?: Skills;
}

export interface PublicPortfolio {
  id: number;
  username: string;
  templateId: string;
  profile?: Profile;
  projects?: Project[];
  skills?: Skills;
}

export interface ProfileUpdate {
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  templateId?: string | null;
  profile?: Profile;
}

export interface ProjectsUpdate {
  projects: Project[];
}

export interface SkillsUpdate {
  skills: Skills;
}

export interface TemplateUpdate {
  templateId: string;
}

export interface ContactInput {
  senderName: string;
  senderEmail: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
}