# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeFolio API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: portfolio
    description: Portfolio management (protected)
  - name: public
    description: Public portfolio access

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/me:
    get:
      operationId: getMe
      tags: [portfolio]
      summary: Get current user portfolio data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/profile:
    put:
      operationId: updateProfile
      tags: [portfolio]
      summary: Update profile fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/projects:
    put:
      operationId: updateProjects
      tags: [portfolio]
      summary: Replace projects array
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectsUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/skills:
    put:
      operationId: updateSkills
      tags: [portfolio]
      summary: Replace skills object
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillsUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/template:
    put:
      operationId: updateTemplate
      tags: [portfolio]
      summary: Set active template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/contact/{username}:
    post:
      operationId: sendContact
      tags: [portfolio]
      summary: Send contact message to a portfolio owner
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /public/{username}:
    get:
      operationId: getPublicPortfolio
      tags: [public]
      summary: Get a public portfolio by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public portfolio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPortfolio"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterInput:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthUser:
      type: object
      required:
        - username
        - email
      properties:
        username:
          type: string
        email:
          type: string

    AuthResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"

    SocialLinks:
      type: object
      properties:
        github:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        website:
          type: ["string", "null"]

    Profile:
      type: object
      properties:
        name:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"

    Project:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        repoLink:
          type: ["string", "null"]
        liveLink:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]

    Skills:
      type: object
      properties:
        frontend:
          type: array
          items:
            type: string
        backend:
          type: array
          items:
            type: string
        devops:
          type: array
          items:
            type: string
        other:
          type: array
          items:
            type: string

    PortfolioUser:
      type: object
      required:
        - id
        - username
        - email
        - templateId
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        templateId:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        skills:
          $ref: "#/components/schemas/Skills"

    PublicPortfolio:
      type: object
      required:
        - id
        - username
        - templateId
      properties:
        id:
          type: integer
        username:
          type: string
        templateId:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        skills:
          $ref: "#/components/schemas/Skills"

    ProfileUpdate:
      type: object
      properties:
        contactEmail:
          type: ["string", "null"]
        templateId:
          type: ["string", "null"]
        profile:
          $ref: "#/components/schemas/Profile"

    ProjectsUpdate:
      type: object
      required:
        - projects
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"

    SkillsUpdate:
      type: object
      required:
        - skills
      properties:
        skills:
          $ref: "#/components/schemas/Skills"

    TemplateUpdate:
      type: object
      required:
        - templateId
      properties:
        templateId:
          type: string

    ContactInput:
      type: object
      required:
        - senderName
        - senderEmail
        - message
      properties:
        senderName:
          type: string
        senderEmail:
          type: string
          format: email
        message:
          type: string

    ContactResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean