# lib/api-zod/src/generated/types/portfolioUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';
import type { Project } from './project';
import type { Skills } from './skills';

export interface PortfolioUser {
  id: number;
  username: string;
  email: string;
  templateId: string;
  profile?: Profile;
  projects?: Project[];
  skills?: Skills;
}