# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from './socialLinks';

export interface Profile {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  domain?: string | null;
  socialLinks?: SocialLinks;
}