# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';

export interface ProfileUpdate {
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  templateId?: string | null;
  profile?: Profile;
}