# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  techStack?: string[];
  /** @nullable */
  repoLink?: string | null;
  /** @nullable */
  liveLink?: string | null;
  /** @nullable */
  screenshotUrl?: string | null;
}