# lib/api-zod/src/generated/types/projectsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';

export interface ProjectsUpdate {
  projects: Project[];
}