# lib/api-zod/src/generated/types/publicPortfolio.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';
import type { Project } from './project';
import type { Skills } from './skills';

export interface PublicPortfolio {
  id: number;
  username: string;
  templateId: string;
  profile?: Profile;
  projects?: Project[];
  skills?: Skills;
}