# lib/api-zod/src/generated/types/skills.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skills {
  frontend?: string[];
  backend?: string[];
  devops?: string[];
  other?: string[];
}