# lib/api-zod/src/generated/types/socialLinks.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLinks {
  /** @nullable */
  github?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  website?: string | null;
}